refactor: Rewrite minio package, modify ffmpeg methods
