+++ pkg/minioX/utils.go
@@ -1,4 +1,4 @@
-package minio
+package minioX
 
 import (
 	"context"
@@ -10,14 +10,24 @@ import (
 	"github.com/minio/minio-go/v7"
 )
 
+type Client struct {
+	Conn *minio.Client
+}
+
+func NewClient(conn *minio.Client) *Client {
+	return &Client{
+		Conn: conn,
+	}
+}
+
 // UploadLocalFile 将本地文件上传至minio
-func UploadLocalFile(filePath string, bucketName string, fileName string, opt minio.PutObjectOptions) error {
+func (c *Client) UploadLocalFile(ctx context.Context, filePath string, bucketName string, fileName string, opt minio.PutObjectOptions) error {
 	// check whether the bucket exists
-	if exists, err := Client.BucketExists(context.Background(), bucketName); !(err == nil && exists) {
+	if exists, err := c.Conn.BucketExists(ctx, bucketName); !(err == nil && exists) {
 		return fmt.Errorf("minio buckect %s miss,err: %v\n", bucketName, err)
 	}
-	uploadInfo, err := Client.FPutObject(
-		context.Background(), bucketName, fileName, filePath, opt)
+	uploadInfo, err := c.Conn.FPutObject(
+		ctx, bucketName, fileName, filePath, opt)
 	if err != nil {
 		return fmt.Errorf("failed uploaded object, err : %w", err)
 	}
@@ -26,11 +36,11 @@ func UploadLocalFile(filePath string, bucketName string, fileName string, opt mi
 }
 
 // UploadSizeFile 读取固定大小的文件并上传至minio(主要使用)
-func UploadSizeFile(bucketName string, fileName string, reader io.Reader, size int64, opt minio.PutObjectOptions) error {
-	if exists, err := Client.BucketExists(context.Background(), bucketName); !(err == nil && exists) {
+func (c *Client) UploadSizeFile(ctx context.Context, bucketName string, fileName string, reader io.Reader, size int64, opt minio.PutObjectOptions) error {
+	if exists, err := c.Conn.BucketExists(ctx, bucketName); !(err == nil && exists) {
 		return fmt.Errorf("minio buckect %s miss,err: %v\n", bucketName, err)
 	}
-	uploadInfo, err := Client.PutObject(context.Background(), bucketName, fileName, reader, size, opt)
+	uploadInfo, err := c.Conn.PutObject(ctx, bucketName, fileName, reader, size, opt)
 	if err != nil {
 		return fmt.Errorf("failed uploaded bytes, err : %w", err)
 	}
@@ -39,14 +49,14 @@ func UploadSizeFile(bucketName string, fileName string, reader io.Reader, size i
 }
 
 // GetFileURL 根据文件名从minio获取文件URL
-func GetFileURL(bucketName string, fileName string, timeLimit time.Duration) (*url.URL, error) {
-	if exists, err := Client.BucketExists(context.Background(), bucketName); !(err == nil && exists) {
+func (c *Client) GetFileURL(ctx context.Context, bucketName string, fileName string, timeLimit time.Duration) (*url.URL, error) {
+	if exists, err := c.Conn.BucketExists(ctx, bucketName); !(err == nil && exists) {
 		return nil, fmt.Errorf("minio buckect %s miss,err: %v\n", bucketName, err)
 	}
 	reqParams := make(url.Values)
 	reqParams.Set("response-content", "attachment; filename=\""+fileName+"\"")
 
-	presignedURL, err := Client.PresignedGetObject(context.Background(), bucketName, fileName, timeLimit, reqParams)
+	presignedURL, err := c.Conn.PresignedGetObject(ctx, bucketName, fileName, timeLimit, reqParams)
 	if err != nil {
 		return nil, fmt.Errorf("failed generated presigned URL, err : %w", err)
 	}
